ddl: migrate test-infra to testify for add index tests
