+++ ddl/db_test.go
@@ -24,7 +24,6 @@ import (
 	"strconv"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"testing"
 	"time"
 
@@ -171,42 +170,6 @@ type testDBSuite8 struct{ *testDBSuite }
 type testSerialDBSuite struct{ *testDBSuite }
 type testSerialDBSuite1 struct{ *testDBSuite }
 
-func testAddIndexWithPK(tk *testkit.TestKit) {
-	tk.MustExec("drop table if exists test_add_index_with_pk")
-	tk.MustExec("create table test_add_index_with_pk(a int not null, b int not null default '0', primary key(a))")
-	tk.MustExec("insert into test_add_index_with_pk values(1, 2)")
-	tk.MustExec("alter table test_add_index_with_pk add index idx (a)")
-	tk.MustQuery("select a from test_add_index_with_pk").Check(testkit.Rows("1"))
-	tk.MustExec("insert into test_add_index_with_pk values(2, 2)")
-	tk.MustExec("alter table test_add_index_with_pk add index idx1 (a, b)")
-	tk.MustQuery("select * from test_add_index_with_pk").Check(testkit.Rows("1 2", "2 2"))
-	tk.MustExec("drop table if exists test_add_index_with_pk1")
-	tk.MustExec("create table test_add_index_with_pk1(a int not null, b int not null default '0', c int, d int, primary key(c))")
-	tk.MustExec("insert into test_add_index_with_pk1 values(1, 1, 1, 1)")
-	tk.MustExec("alter table test_add_index_with_pk1 add index idx (c)")
-	tk.MustExec("insert into test_add_index_with_pk1 values(2, 2, 2, 2)")
-	tk.MustQuery("select * from test_add_index_with_pk1").Check(testkit.Rows("1 1 1 1", "2 2 2 2"))
-	tk.MustExec("drop table if exists test_add_index_with_pk2")
-	tk.MustExec("create table test_add_index_with_pk2(a int not null, b int not null default '0', c int unsigned, d int, primary key(c))")
-	tk.MustExec("insert into test_add_index_with_pk2 values(1, 1, 1, 1)")
-	tk.MustExec("alter table test_add_index_with_pk2 add index idx (c)")
-	tk.MustExec("insert into test_add_index_with_pk2 values(2, 2, 2, 2)")
-	tk.MustQuery("select * from test_add_index_with_pk2").Check(testkit.Rows("1 1 1 1", "2 2 2 2"))
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (a int, b int, c int, primary key(a, b));")
-	tk.MustExec("insert into t values (1, 2, 3);")
-	tk.MustExec("create index idx on t (a, b);")
-}
-
-func (s *testDBSuite7) TestAddIndexWithPK(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use " + s.schemaName)
-
-	testAddIndexWithPK(tk)
-	tk.Se.GetSessionVars().EnableClusteredIndex = variable.ClusteredIndexDefModeOn
-	testAddIndexWithPK(tk)
-}
-
 func (s *testDBSuite5) TestAddIndexWithDupIndex(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use " + s.schemaName)
@@ -323,32 +286,6 @@ func backgroundExecT(s kv.Storage, sql string, done chan error) {
 	}
 	_, err = se.Execute(context.Background(), sql)
 	done <- errors.Trace(err)
-}
-
-// TestAddPrimaryKeyRollback1 is used to test scenarios that will roll back when a duplicate primary key is encountered.
-func (s *testDBSuite8) TestAddPrimaryKeyRollback1(c *C) {
-	hasNullValsInKey := false
-	idxName := "PRIMARY"
-	addIdxSQL := "alter table t1 add primary key c3_index (c3);"
-	errMsg := "[kv:1062]Duplicate entry '" + strconv.Itoa(defaultBatchSize*2-10) + "' for key 'PRIMARY'"
-	testAddIndexRollback(c, s.store, s.lease, idxName, addIdxSQL, errMsg, hasNullValsInKey)
-}
-
-// TestAddPrimaryKeyRollback2 is used to test scenarios that will roll back when a null primary key is encountered.
-func (s *testDBSuite8) TestAddPrimaryKeyRollback2(c *C) {
-	hasNullValsInKey := true
-	idxName := "PRIMARY"
-	addIdxSQL := "alter table t1 add primary key c3_index (c3);"
-	errMsg := "[ddl:1138]Invalid use of NULL value"
-	testAddIndexRollback(c, s.store, s.lease, idxName, addIdxSQL, errMsg, hasNullValsInKey)
-}
-
-func (s *testDBSuite2) TestAddUniqueIndexRollback(c *C) {
-	hasNullValsInKey := false
-	idxName := "c3_index"
-	addIdxSQL := "create unique index c3_index on t1 (c3)"
-	errMsg := "[kv:1062]Duplicate entry '" + strconv.Itoa(defaultBatchSize*2-10) + "' for key 'c3_index'"
-	testAddIndexRollback(c, s.store, s.lease, idxName, addIdxSQL, errMsg, hasNullValsInKey)
 }
 
 func (s *testSerialDBSuite) TestWriteReorgForColumnTypeChangeOnAmendTxn(c *C) {
@@ -514,303 +451,6 @@ func batchInsertLegacy(tk *testkit.TestKit, tbl string, start, end int) {
 		}
 	}
 	tk.MustExec(dml)
-}
-
-func testAddIndexRollback(c *C, store kv.Storage, lease time.Duration, idxName, addIdxSQL, errMsg string, hasNullValsInKey bool) {
-	tk := testkit.NewTestKit(c, store)
-	tk.MustExec("use test_db")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("create table t1 (c1 int, c2 int, c3 int, unique key(c1))")
-	// defaultBatchSize is equal to ddl.defaultBatchSize
-	base := defaultBatchSize * 2
-	count := base
-	// add some rows
-	batchInsertLegacy(tk, "t1", 0, count)
-	// add some null rows
-	if hasNullValsInKey {
-		for i := count - 10; i < count; i++ {
-			tk.MustExec("insert into t1 values (?, ?, null)", i+10, i)
-		}
-	} else {
-		// add some duplicate rows
-		for i := count - 10; i < count; i++ {
-			tk.MustExec("insert into t1 values (?, ?, ?)", i+10, i, i)
-		}
-	}
-
-	done := make(chan error, 1)
-	go backgroundExec(store, addIdxSQL, done)
-
-	times := 0
-	ticker := time.NewTicker(lease / 2)
-	defer ticker.Stop()
-LOOP:
-	for {
-		select {
-		case err := <-done:
-			c.Assert(err, NotNil)
-			c.Assert(err.Error(), Equals, errMsg, Commentf("err:%v", err))
-			break LOOP
-		case <-ticker.C:
-			if times >= 10 {
-				break
-			}
-			step := 5
-			// delete some rows, and add some data
-			for i := count; i < count+step; i++ {
-				n := rand.Intn(count)
-				// (2048, 2038, 2038) and (2038, 2038, 2038)
-				// Don't delete rows where c1 is 2048 or 2038, otherwise, the entry value in duplicated error message would change.
-				if n == defaultBatchSize*2-10 || n == defaultBatchSize*2 {
-					continue
-				}
-				tk.MustExec("delete from t1 where c1 = ?", n)
-				tk.MustExec("insert into t1 values (?, ?, ?)", i+10, i, i)
-			}
-			count += step
-			times++
-		}
-	}
-
-	ctx := tk.Se.(sessionctx.Context)
-	t := testGetTableByName(c, ctx, "test_db", "t1")
-	for _, tidx := range t.Indices() {
-		c.Assert(strings.EqualFold(tidx.Meta().Name.L, idxName), IsFalse)
-	}
-
-	// delete duplicated/null rows, then add index
-	for i := base - 10; i < base; i++ {
-		tk.MustExec("delete from t1 where c1 = ?", i+10)
-	}
-	sessionExec(c, store, addIdxSQL)
-	tk.MustExec("drop table t1")
-}
-
-func (s *testDBSuite8) TestCancelAddPrimaryKey(c *C) {
-	idxName := "primary"
-	addIdxSQL := "alter table t1 add primary key idx_c2 (c2);"
-	testCancelAddIndex(c, s.store, s.dom.DDL(), s.lease, idxName, addIdxSQL, "", s.dom)
-
-	// Check the column's flag when the "add primary key" failed.
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test_db")
-	ctx := tk.Se.(sessionctx.Context)
-	c.Assert(ctx.NewTxn(context.Background()), IsNil)
-	t := testGetTableByName(c, ctx, "test_db", "t1")
-	col1Flag := t.Cols()[1].Flag
-	c.Assert(!mysql.HasNotNullFlag(col1Flag) && !mysql.HasPreventNullInsertFlag(col1Flag) && mysql.HasUnsignedFlag(col1Flag), IsTrue)
-	tk.MustExec("drop table t1")
-}
-
-func (s *testDBSuite7) TestCancelAddIndex(c *C) {
-	idxName := "c3_index"
-	addIdxSQL := "create unique index c3_index on t1 (c3)"
-	testCancelAddIndex(c, s.store, s.dom.DDL(), s.lease, idxName, addIdxSQL, "", s.dom)
-
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test_db")
-	tk.MustExec("drop table t1")
-}
-
-func backgroundExecOnJobUpdatedExported(c *C, store kv.Storage, ctx sessionctx.Context, hook *ddl.TestDDLCallback, idxName string) (
-	func(*model.Job), *model.IndexInfo, error) {
-	var checkErr error
-	first := true
-	c3IdxInfo := &model.IndexInfo{}
-	hook.OnJobUpdatedExported = func(job *model.Job) {
-		addIndexNotFirstReorg := (job.Type == model.ActionAddIndex || job.Type == model.ActionAddPrimaryKey) &&
-			job.SchemaState == model.StateWriteReorganization && job.SnapshotVer != 0
-		// If the action is adding index and the state is writing reorganization, it want to test the case of cancelling the job when backfilling indexes.
-		// When the job satisfies this case of addIndexNotFirstReorg, the worker will start to backfill indexes.
-		if !addIndexNotFirstReorg {
-			// Get the index's meta.
-			if c3IdxInfo.ID != 0 {
-				return
-			}
-			t := testGetTableByName(c, ctx, "test_db", "t1")
-			for _, index := range t.Indices() {
-				if !tables.IsIndexWritable(index) {
-					continue
-				}
-				if index.Meta().Name.L == idxName {
-					*c3IdxInfo = *index.Meta()
-				}
-			}
-			return
-		}
-		// The job satisfies the case of addIndexNotFirst for the first time, the worker hasn't finished a batch of backfill indexes.
-		if first {
-			first = false
-			return
-		}
-		if checkErr != nil {
-			return
-		}
-		hookCtx := mock.NewContext()
-		hookCtx.Store = store
-		err := hookCtx.NewTxn(context.Background())
-		if err != nil {
-			checkErr = errors.Trace(err)
-			return
-		}
-		jobIDs := []int64{job.ID}
-		txn, err := hookCtx.Txn(true)
-		if err != nil {
-			checkErr = errors.Trace(err)
-			return
-		}
-		errs, err := admin.CancelJobs(txn, jobIDs)
-		if err != nil {
-			checkErr = errors.Trace(err)
-			return
-		}
-		// It only tests cancel one DDL job.
-		if errs[0] != nil {
-			checkErr = errors.Trace(errs[0])
-			return
-		}
-		txn, err = hookCtx.Txn(true)
-		if err != nil {
-			checkErr = errors.Trace(err)
-			return
-		}
-		err = txn.Commit(context.Background())
-		if err != nil {
-			checkErr = errors.Trace(err)
-		}
-	}
-	return hook.OnJobUpdatedExported, c3IdxInfo, checkErr
-}
-
-func testCancelAddIndex(c *C, store kv.Storage, d ddl.DDL, lease time.Duration, idxName, addIdxSQL, sqlModeSQL string, dom *domain.Domain) {
-	tk := testkit.NewTestKit(c, store)
-	tk.MustExec("use test_db")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("create table t1 (c1 int, c2 int unsigned, c3 int, unique key(c1))")
-	// defaultBatchSize is equal to ddl.defaultBatchSize
-	count := defaultBatchSize * 32
-	start := 0
-	// add some rows
-	if len(sqlModeSQL) != 0 {
-		// Insert some null values.
-		tk.MustExec(sqlModeSQL)
-		tk.MustExec("insert into t1 set c1 = ?", 0)
-		tk.MustExec("insert into t1 set c2 = ?", 1)
-		tk.MustExec("insert into t1 set c3 = ?", 2)
-		start = 3
-	}
-	for i := start; i < count; i += defaultBatchSize {
-		batchInsertLegacy(tk, "t1", i, i+defaultBatchSize)
-	}
-
-	var c3IdxInfo *model.IndexInfo
-	hook := &ddl.TestDDLCallback{Do: dom}
-	originBatchSize := tk.MustQuery("select @@global.tidb_ddl_reorg_batch_size")
-	// Set batch size to lower try to slow down add-index reorganization, This if for hook to cancel this ddl job.
-	tk.MustExec("set @@global.tidb_ddl_reorg_batch_size = 32")
-	defer tk.MustExec(fmt.Sprintf("set @@global.tidb_ddl_reorg_batch_size = %v", originBatchSize.Rows()[0][0]))
-	// let hook.OnJobUpdatedExported has chance to cancel the job.
-	// the hook.OnJobUpdatedExported is called when the job is updated, runReorgJob will wait ddl.ReorgWaitTimeout, then return the ddl.runDDLJob.
-	// After that ddl call d.hook.OnJobUpdated(job), so that we can canceled the job in this test case.
-	var checkErr error
-	ctx := tk.Se.(sessionctx.Context)
-	hook.OnJobUpdatedExported, c3IdxInfo, checkErr = backgroundExecOnJobUpdatedExported(c, store, ctx, hook, idxName)
-	originalHook := d.GetHook()
-	jobIDExt := wrapJobIDExtCallback(hook)
-	d.SetHook(jobIDExt)
-	done := make(chan error, 1)
-	go backgroundExec(store, addIdxSQL, done)
-
-	times := 0
-	ticker := time.NewTicker(lease / 2)
-	defer ticker.Stop()
-LOOP:
-	for {
-		select {
-		case err := <-done:
-			c.Assert(checkErr, IsNil)
-			c.Assert(err, NotNil)
-			c.Assert(err.Error(), Equals, "[ddl:8214]Cancelled DDL job")
-			break LOOP
-		case <-ticker.C:
-			if times >= 10 {
-				break
-			}
-			step := 5
-			// delete some rows, and add some data
-			for i := count; i < count+step; i++ {
-				n := rand.Intn(count)
-				tk.MustExec("delete from t1 where c1 = ?", n)
-				tk.MustExec("insert into t1 values (?, ?, ?)", i+10, i, i)
-			}
-			count += step
-			times++
-		}
-	}
-	checkDelRangeAdded(tk, jobIDExt.jobID, c3IdxInfo.ID)
-	d.SetHook(originalHook)
-}
-
-// TestCancelAddIndex1 tests canceling ddl job when the add index worker is not started.
-func (s *testDBSuite4) TestCancelAddIndex1(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.mustExec(tk, c, "use test_db")
-	s.mustExec(tk, c, "drop table if exists t")
-	s.mustExec(tk, c, "create table t(c1 int, c2 int)")
-	defer s.mustExec(tk, c, "drop table t;")
-
-	for i := 0; i < 50; i++ {
-		s.mustExec(tk, c, "insert into t values (?, ?)", i, i)
-	}
-
-	var checkErr error
-	hook := &ddl.TestDDLCallback{Do: s.dom}
-	hook.OnJobRunBeforeExported = func(job *model.Job) {
-		if job.Type == model.ActionAddIndex && job.State == model.JobStateRunning && job.SchemaState == model.StateWriteReorganization && job.SnapshotVer == 0 {
-			jobIDs := []int64{job.ID}
-			hookCtx := mock.NewContext()
-			hookCtx.Store = s.store
-			err := hookCtx.NewTxn(context.Background())
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-			txn, err := hookCtx.Txn(true)
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-			errs, err := admin.CancelJobs(txn, jobIDs)
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-
-			if errs[0] != nil {
-				checkErr = errors.Trace(errs[0])
-				return
-			}
-
-			checkErr = txn.Commit(context.Background())
-		}
-	}
-	originalHook := s.dom.DDL().GetHook()
-	s.dom.DDL().SetHook(hook)
-	rs, err := tk.Exec("alter table t add index idx_c2(c2)")
-	if rs != nil {
-		rs.Close()
-	}
-	c.Assert(checkErr, IsNil)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "[ddl:8214]Cancelled DDL job")
-
-	s.dom.DDL().SetHook(originalHook)
-	t := s.testGetTable(c, "t")
-	for _, idx := range t.Indices() {
-		c.Assert(strings.EqualFold(idx.Meta().Name.L, "idx_c2"), IsFalse)
-	}
-	s.mustExec(tk, c, "alter table t add index idx_c2(c2)")
-	s.mustExec(tk, c, "alter table t drop index idx_c2")
 }
 
 // TestCancelDropIndex tests cancel ddl job which type is drop primary key.
@@ -1171,65 +811,6 @@ func (s *testDBSuite2) TestCancelDropTableAndSchema(c *C) {
 	}
 }
 
-func (s *testDBSuite3) TestAddAnonymousIndex(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use " + s.schemaName)
-	s.mustExec(tk, c, "create table t_anonymous_index (c1 int, c2 int, C3 int)")
-	s.mustExec(tk, c, "alter table t_anonymous_index add index (c1, c2)")
-	// for dropping empty index
-	_, err := tk.Exec("alter table t_anonymous_index drop index")
-	c.Assert(err, NotNil)
-	// The index name is c1 when adding index (c1, c2).
-	s.mustExec(tk, c, "alter table t_anonymous_index drop index c1")
-	t := s.testGetTable(c, "t_anonymous_index")
-	c.Assert(t.Indices(), HasLen, 0)
-	// for adding some indices that the first column name is c1
-	s.mustExec(tk, c, "alter table t_anonymous_index add index (c1)")
-	_, err = tk.Exec("alter table t_anonymous_index add index c1 (c2)")
-	c.Assert(err, NotNil)
-	t = s.testGetTable(c, "t_anonymous_index")
-	c.Assert(t.Indices(), HasLen, 1)
-	idx := t.Indices()[0].Meta().Name.L
-	c.Assert(idx, Equals, "c1")
-	// The MySQL will be a warning.
-	s.mustExec(tk, c, "alter table t_anonymous_index add index c1_3 (c1)")
-	s.mustExec(tk, c, "alter table t_anonymous_index add index (c1, c2, C3)")
-	// The MySQL will be a warning.
-	s.mustExec(tk, c, "alter table t_anonymous_index add index (c1)")
-	t = s.testGetTable(c, "t_anonymous_index")
-	c.Assert(t.Indices(), HasLen, 4)
-	s.mustExec(tk, c, "alter table t_anonymous_index drop index c1")
-	s.mustExec(tk, c, "alter table t_anonymous_index drop index c1_2")
-	s.mustExec(tk, c, "alter table t_anonymous_index drop index c1_3")
-	s.mustExec(tk, c, "alter table t_anonymous_index drop index c1_4")
-	// for case insensitive
-	s.mustExec(tk, c, "alter table t_anonymous_index add index (C3)")
-	s.mustExec(tk, c, "alter table t_anonymous_index drop index c3")
-	s.mustExec(tk, c, "alter table t_anonymous_index add index c3 (C3)")
-	s.mustExec(tk, c, "alter table t_anonymous_index drop index C3")
-	// for anonymous index with column name `primary`
-	s.mustExec(tk, c, "create table t_primary (`primary` int, b int, key (`primary`))")
-	t = s.testGetTable(c, "t_primary")
-	c.Assert(t.Indices()[0].Meta().Name.String(), Equals, "primary_2")
-	s.mustExec(tk, c, "alter table t_primary add index (`primary`);")
-	t = s.testGetTable(c, "t_primary")
-	c.Assert(t.Indices()[0].Meta().Name.String(), Equals, "primary_2")
-	c.Assert(t.Indices()[1].Meta().Name.String(), Equals, "primary_3")
-	s.mustExec(tk, c, "alter table t_primary add primary key(b);")
-	t = s.testGetTable(c, "t_primary")
-	c.Assert(t.Indices()[0].Meta().Name.String(), Equals, "primary_2")
-	c.Assert(t.Indices()[1].Meta().Name.String(), Equals, "primary_3")
-	c.Assert(t.Indices()[2].Meta().Name.L, Equals, "primary")
-	s.mustExec(tk, c, "create table t_primary_2 (`primary` int, key primary_2 (`primary`), key (`primary`))")
-	t = s.testGetTable(c, "t_primary_2")
-	c.Assert(t.Indices()[0].Meta().Name.String(), Equals, "primary_2")
-	c.Assert(t.Indices()[1].Meta().Name.String(), Equals, "primary_3")
-	s.mustExec(tk, c, "create table t_primary_3 (`primary_2` int, key(`primary_2`), `primary` int, key(`primary`));")
-	t = s.testGetTable(c, "t_primary_3")
-	c.Assert(t.Indices()[0].Meta().Name.String(), Equals, "primary_2")
-	c.Assert(t.Indices()[1].Meta().Name.String(), Equals, "primary_3")
-}
-
 func (s *testDBSuite4) TestAlterLock(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use " + s.schemaName)
@@ -1276,393 +857,6 @@ func (s *testDBSuite6) TestAddMultiColumnsIndexClusterIndex(c *C) {
 
 	tk.MustExec("admin check index t idx;")
 	tk.MustExec("admin check table t;")
-}
-
-func (s *testDBSuite6) TestAddPrimaryKey1(c *C) {
-	testAddIndex(c, s.store, s.lease, testPlain,
-		"create table test_add_index (c1 bigint, c2 bigint, c3 bigint, unique key(c1))", "primary")
-}
-
-func (s *testDBSuite2) TestAddPrimaryKey2(c *C) {
-	testAddIndex(c, s.store, s.lease, testPartition,
-		`create table test_add_index (c1 bigint, c2 bigint, c3 bigint, key(c1))
-			      partition by range (c3) (
-			      partition p0 values less than (3440),
-			      partition p1 values less than (61440),
-			      partition p2 values less than (122880),
-			      partition p3 values less than (204800),
-			      partition p4 values less than maxvalue)`, "primary")
-}
-
-func (s *testDBSuite3) TestAddPrimaryKey3(c *C) {
-	testAddIndex(c, s.store, s.lease, testPartition,
-		`create table test_add_index (c1 bigint, c2 bigint, c3 bigint, key(c1))
-			      partition by hash (c3) partitions 4;`, "primary")
-}
-
-func (s *testDBSuite4) TestAddPrimaryKey4(c *C) {
-	testAddIndex(c, s.store, s.lease, testPartition,
-		`create table test_add_index (c1 bigint, c2 bigint, c3 bigint, key(c1))
-			      partition by range columns (c3) (
-			      partition p0 values less than (3440),
-			      partition p1 values less than (61440),
-			      partition p2 values less than (122880),
-			      partition p3 values less than (204800),
-			      partition p4 values less than maxvalue)`, "primary")
-}
-
-func (s *testDBSuite6) TestAddIndex1(c *C) {
-	testAddIndex(c, s.store, s.lease, testPlain,
-		"create table test_add_index (c1 bigint, c2 bigint, c3 bigint, primary key(c1))", "")
-}
-
-func (s *testSerialDBSuite1) TestAddIndex1WithShardRowID(c *C) {
-	testAddIndex(c, s.store, s.lease, testPartition|testShardRowID,
-		"create table test_add_index (c1 bigint, c2 bigint, c3 bigint) SHARD_ROW_ID_BITS = 4 pre_split_regions = 4;", "")
-}
-
-func (s *testDBSuite2) TestAddIndex2(c *C) {
-	testAddIndex(c, s.store, s.lease, testPartition,
-		`create table test_add_index (c1 bigint, c2 bigint, c3 bigint, primary key(c1))
-			      partition by range (c1) (
-			      partition p0 values less than (3440),
-			      partition p1 values less than (61440),
-			      partition p2 values less than (122880),
-			      partition p3 values less than (204800),
-			      partition p4 values less than maxvalue)`, "")
-}
-
-func (s *testSerialDBSuite) TestAddIndex2WithShardRowID(c *C) {
-	testAddIndex(c, s.store, s.lease, testPartition|testShardRowID,
-		`create table test_add_index (c1 bigint, c2 bigint, c3 bigint)
-				  SHARD_ROW_ID_BITS = 4 pre_split_regions = 4
-			      partition by range (c1) (
-			      partition p0 values less than (3440),
-			      partition p1 values less than (61440),
-			      partition p2 values less than (122880),
-			      partition p3 values less than (204800),
-			      partition p4 values less than maxvalue)`, "")
-}
-
-func (s *testDBSuite3) TestAddIndex3(c *C) {
-	testAddIndex(c, s.store, s.lease, testPartition,
-		`create table test_add_index (c1 bigint, c2 bigint, c3 bigint, primary key(c1))
-			      partition by hash (c1) partitions 4;`, "")
-}
-
-func (s *testSerialDBSuite1) TestAddIndex3WithShardRowID(c *C) {
-	testAddIndex(c, s.store, s.lease, testPartition|testShardRowID,
-		`create table test_add_index (c1 bigint, c2 bigint, c3 bigint)
-				  SHARD_ROW_ID_BITS = 4 pre_split_regions = 4
-			      partition by hash (c1) partitions 4;`, "")
-}
-
-func (s *testDBSuite8) TestAddIndex4(c *C) {
-	testAddIndex(c, s.store, s.lease, testPartition,
-		`create table test_add_index (c1 bigint, c2 bigint, c3 bigint, primary key(c1))
-			      partition by range columns (c1) (
-			      partition p0 values less than (3440),
-			      partition p1 values less than (61440),
-			      partition p2 values less than (122880),
-			      partition p3 values less than (204800),
-			      partition p4 values less than maxvalue)`, "")
-}
-
-func (s *testSerialDBSuite) TestAddIndex4WithShardRowID(c *C) {
-	testAddIndex(c, s.store, s.lease, testPartition|testShardRowID,
-		`create table test_add_index (c1 bigint, c2 bigint, c3 bigint)
-				  SHARD_ROW_ID_BITS = 4 pre_split_regions = 4
-			      partition by range columns (c1) (
-			      partition p0 values less than (3440),
-			      partition p1 values less than (61440),
-			      partition p2 values less than (122880),
-			      partition p3 values less than (204800),
-			      partition p4 values less than maxvalue)`, "")
-}
-
-func (s *testDBSuite5) TestAddIndex5(c *C) {
-	testAddIndex(c, s.store, s.lease, testClusteredIndex,
-		`create table test_add_index (c1 bigint, c2 bigint, c3 bigint, primary key(c2, c3))`, "")
-}
-
-type testAddIndexType uint8
-
-const (
-	testPlain          testAddIndexType = 1
-	testPartition      testAddIndexType = 1 << 1
-	testClusteredIndex testAddIndexType = 1 << 2
-	testShardRowID     testAddIndexType = 1 << 3
-)
-
-func testAddIndex(c *C, store kv.Storage, lease time.Duration, tp testAddIndexType, createTableSQL, idxTp string) {
-	tk := testkit.NewTestKit(c, store)
-	tk.MustExec("use test_db")
-	isTestPartition := (testPartition & tp) > 0
-	isTestShardRowID := (testShardRowID & tp) > 0
-	if isTestShardRowID {
-		atomic.StoreUint32(&ddl.EnableSplitTableRegion, 1)
-		tk.MustExec("set global tidb_scatter_region = 1")
-		defer func() {
-			atomic.StoreUint32(&ddl.EnableSplitTableRegion, 0)
-			tk.MustExec("set global tidb_scatter_region = 0")
-		}()
-	}
-	if isTestPartition {
-		tk.MustExec("set @@session.tidb_enable_table_partition = '1';")
-	} else if (testClusteredIndex & tp) > 0 {
-		tk.Se.GetSessionVars().EnableClusteredIndex = variable.ClusteredIndexDefModeOn
-	}
-	tk.MustExec("drop table if exists test_add_index")
-	tk.MustExec(createTableSQL)
-
-	done := make(chan error, 1)
-	start := -10
-	num := defaultBatchSize
-	// first add some rows
-	batchInsertLegacy(tk, "test_add_index", start, num)
-
-	// Add some discrete rows.
-	maxBatch := 20
-	batchCnt := 100
-	otherKeys := make([]int, 0, batchCnt*maxBatch)
-	// Make sure there are no duplicate keys.
-	base := defaultBatchSize * 20
-	for i := 1; i < batchCnt; i++ {
-		if isTestShardRowID {
-			base = i % 4 << 61
-		}
-		n := base + i*defaultBatchSize + i
-		for j := 0; j < rand.Intn(maxBatch); j++ {
-			n += j
-			sql := fmt.Sprintf("insert into test_add_index values (%d, %d, %d)", n, n, n)
-			tk.MustExec(sql)
-			otherKeys = append(otherKeys, n)
-		}
-	}
-	// Encounter the value of math.MaxInt64 in middle of
-	v := math.MaxInt64 - defaultBatchSize/2
-	tk.MustExec(fmt.Sprintf("insert into test_add_index values (%d, %d, %d)", v, v, v))
-	otherKeys = append(otherKeys, v)
-
-	addIdxSQL := fmt.Sprintf("alter table test_add_index add %s key c3_index(c3)", idxTp)
-	testddlutil.SessionExecInGoroutine(store, addIdxSQL, done)
-
-	deletedKeys := make(map[int]struct{})
-
-	ticker := time.NewTicker(lease / 2)
-	defer ticker.Stop()
-LOOP:
-	for {
-		select {
-		case err := <-done:
-			if err == nil {
-				break LOOP
-			}
-			c.Assert(err, IsNil, Commentf("err:%v", errors.ErrorStack(err)))
-		case <-ticker.C:
-			// When the server performance is particularly poor,
-			// the adding index operation can not be completed.
-			// So here is a limit to the number of rows inserted.
-			if num > defaultBatchSize*10 {
-				break
-			}
-			step := 5
-			// delete some rows, and add some data
-			for i := num; i < num+step; i++ {
-				n := rand.Intn(num)
-				deletedKeys[n] = struct{}{}
-				sql := fmt.Sprintf("delete from test_add_index where c1 = %d", n)
-				tk.MustExec(sql)
-				sql = fmt.Sprintf("insert into test_add_index values (%d, %d, %d)", i, i, i)
-				tk.MustExec(sql)
-			}
-			num += step
-		}
-	}
-
-	if isTestShardRowID {
-		re := tk.MustQuery("show table test_add_index regions;")
-		rows := re.Rows()
-		c.Assert(len(rows), GreaterEqual, 16)
-		tk.MustExec("admin check table test_add_index")
-		return
-	}
-
-	// get exists keys
-	keys := make([]int, 0, num)
-	for i := start; i < num; i++ {
-		if _, ok := deletedKeys[i]; ok {
-			continue
-		}
-		keys = append(keys, i)
-	}
-	keys = append(keys, otherKeys...)
-
-	// test index key
-	expectedRows := make([][]interface{}, 0, len(keys))
-	for _, key := range keys {
-		expectedRows = append(expectedRows, []interface{}{key})
-	}
-	rows := tk.MustQuery(fmt.Sprintf("select c1 from test_add_index where c3 >= %d order by c1", start)).Rows()
-	matchRows(c, rows, expectedRows)
-
-	tk.MustExec("admin check table test_add_index")
-	if isTestPartition {
-		return
-	}
-
-	// TODO: Support explain in future.
-	// rows := s.mustQuery(c, "explain select c1 from test_add_index where c3 >= 100")
-
-	// ay := dumpRows(c, rows)
-	// c.Assert(strings.Contains(fmt.Sprintf("%v", ay), "c3_index"), IsTrue)
-
-	// get all row handles
-	ctx := tk.Se.(sessionctx.Context)
-	c.Assert(ctx.NewTxn(context.Background()), IsNil)
-	t := testGetTableByName(c, ctx, "test_db", "test_add_index")
-	handles := kv.NewHandleMap()
-	err := tables.IterRecords(t, ctx, t.Cols(),
-		func(h kv.Handle, data []types.Datum, cols []*table.Column) (bool, error) {
-			handles.Set(h, struct{}{})
-			return true, nil
-		})
-	c.Assert(err, IsNil)
-
-	// check in index
-	var nidx table.Index
-	idxName := "c3_index"
-	if len(idxTp) != 0 {
-		idxName = "primary"
-	}
-	for _, tidx := range t.Indices() {
-		if tidx.Meta().Name.L == idxName {
-			nidx = tidx
-			break
-		}
-	}
-	// Make sure there is index with name c3_index.
-	c.Assert(nidx, NotNil)
-	c.Assert(nidx.Meta().ID, Greater, int64(0))
-	txn, err := ctx.Txn(true)
-	c.Assert(err, IsNil)
-	err = txn.Rollback()
-	c.Assert(err, IsNil)
-
-	c.Assert(ctx.NewTxn(context.Background()), IsNil)
-
-	tk.MustExec("admin check table test_add_index")
-	tk.MustExec("drop table test_add_index")
-}
-
-func (s *testDBSuite1) TestAddIndexWithSplitTable(c *C) {
-	createSQL := "CREATE TABLE test_add_index(a bigint PRIMARY KEY AUTO_RANDOM(4), b varchar(255), c bigint)"
-	stSQL := fmt.Sprintf("SPLIT TABLE test_add_index BETWEEN (%d) AND (%d) REGIONS 16;", math.MinInt64, math.MaxInt64)
-	testAddIndexWithSplitTable(c, s.store, s.lease, createSQL, stSQL)
-}
-
-func (s *testSerialDBSuite) TestAddIndexWithShardRowID(c *C) {
-	createSQL := "create table test_add_index(a bigint, b bigint, c bigint) SHARD_ROW_ID_BITS = 4 pre_split_regions = 4;"
-	testAddIndexWithSplitTable(c, s.store, s.lease, createSQL, "")
-}
-
-func testAddIndexWithSplitTable(c *C, store kv.Storage, lease time.Duration, createSQL, splitTableSQL string) {
-	tk := testkit.NewTestKit(c, store)
-	tk.MustExec("use test_db")
-	tk.MustExec("drop table if exists test_add_index")
-	hasAutoRadomField := len(splitTableSQL) > 0
-	if !hasAutoRadomField {
-		atomic.StoreUint32(&ddl.EnableSplitTableRegion, 1)
-		tk.MustExec("set global tidb_scatter_region = 1")
-		defer func() {
-			atomic.StoreUint32(&ddl.EnableSplitTableRegion, 0)
-			tk.MustExec("set global tidb_scatter_region = 0")
-		}()
-	}
-	tk.MustExec(createSQL)
-
-	batchInsertRows := func(tk *testkit.TestKit, needVal bool, tbl string, start, end int) error {
-		dml := fmt.Sprintf("insert into %s values", tbl)
-		for i := start; i < end; i++ {
-			if needVal {
-				dml += fmt.Sprintf("(%d, %d, %d)", i, i, i)
-			} else {
-				dml += "()"
-			}
-			if i != end-1 {
-				dml += ","
-			}
-		}
-		_, err := tk.Exec(dml)
-		return err
-	}
-
-	done := make(chan error, 1)
-	start := -20
-	num := defaultBatchSize
-	// Add some discrete rows.
-	goCnt := 10
-	errCh := make(chan error, goCnt)
-	for i := 0; i < goCnt; i++ {
-		base := (i % 8) << 60
-		go func(b int, eCh chan error) {
-			tk1 := testkit.NewTestKit(c, store)
-			tk1.MustExec("use test_db")
-			eCh <- batchInsertRows(tk1, !hasAutoRadomField, "test_add_index", base+start, base+num)
-		}(base, errCh)
-	}
-	for i := 0; i < goCnt; i++ {
-		e := <-errCh
-		c.Assert(e, IsNil)
-	}
-
-	if hasAutoRadomField {
-		tk.MustQuery(splitTableSQL).Check(testkit.Rows("15 1"))
-	}
-	tk.MustQuery("select @@session.tidb_wait_split_region_finish;").Check(testkit.Rows("1"))
-	re := tk.MustQuery("show table test_add_index regions;")
-	rows := re.Rows()
-	c.Assert(len(rows), Equals, 16)
-	addIdxSQL := "alter table test_add_index add index idx(a)"
-	testddlutil.SessionExecInGoroutine(store, addIdxSQL, done)
-
-	ticker := time.NewTicker(lease / 5)
-	defer ticker.Stop()
-	num = 0
-LOOP:
-	for {
-		select {
-		case err := <-done:
-			if err == nil {
-				break LOOP
-			}
-			c.Assert(err, IsNil, Commentf("err:%v", errors.ErrorStack(err)))
-		case <-ticker.C:
-			// When the server performance is particularly poor,
-			// the adding index operation can not be completed.
-			// So here is a limit to the number of rows inserted.
-			if num >= 1000 {
-				break
-			}
-			step := 20
-			// delete, insert and update some data
-			for i := num; i < num+step; i++ {
-				sql := fmt.Sprintf("delete from test_add_index where a = %d", i+1)
-				tk.MustExec(sql)
-				if hasAutoRadomField {
-					sql = "insert into test_add_index values ()"
-				} else {
-					sql = fmt.Sprintf("insert into test_add_index values (%d, %d, %d)", i, i, i)
-				}
-				tk.MustExec(sql)
-				sql = fmt.Sprintf("update test_add_index set b = %d", i*10)
-				tk.MustExec(sql)
-			}
-			num += step
-		}
-	}
-
-	tk.MustExec("admin check table test_add_index")
 }
 
 // TestCancelAddTableAndDropTablePartition tests cancel ddl job which type is add/drop table partition.
@@ -1795,7 +989,7 @@ func testDropIndex(c *C, store kv.Storage, lease time.Duration, createSQL, dropI
 	indexID := testGetIndexID(c, ctx, "test_db", "test_drop_index", idxName)
 	jobIDExt, reset := setupJobIDExtCallback(ctx)
 	defer reset()
-	testddlutil.SessionExecInGoroutine(store, dropIdxSQL, done)
+	testddlutil.SessionExecInGoroutine(store, "test_db", dropIdxSQL, done)
 
 	ticker := time.NewTicker(lease / 2)
 	defer ticker.Stop()
@@ -2376,17 +1570,6 @@ func (s *testDBSuite6) TestColumn(c *C) {
 	tk.MustExec("drop table t2")
 }
 
-func sessionExec(c *C, s kv.Storage, sql string) {
-	se, err := session.CreateSession4Test(s)
-	c.Assert(err, IsNil)
-	_, err = se.Execute(context.Background(), "use test_db")
-	c.Assert(err, IsNil)
-	rs, err := se.Execute(context.Background(), sql)
-	c.Assert(err, IsNil, Commentf("err:%v", errors.ErrorStack(err)))
-	c.Assert(rs, IsNil)
-	se.Close()
-}
-
 func (s *testDBSuite) testAddColumn(tk *testkit.TestKit, c *C) {
 	done := make(chan error, 1)
 
@@ -2394,7 +1577,7 @@ func (s *testDBSuite) testAddColumn(tk *testkit.TestKit, c *C) {
 	// add some rows
 	batchInsertLegacy(tk, "t2", 0, num)
 
-	testddlutil.SessionExecInGoroutine(s.store, "alter table t2 add column c4 int default -1", done)
+	testddlutil.SessionExecInGoroutine(s.store, "test_db", "alter table t2 add column c4 int default -1", done)
 
 	ticker := time.NewTicker(s.lease / 2)
 	defer ticker.Stop()
@@ -2540,7 +1723,7 @@ func (s *testDBSuite) testDropColumn(tk *testkit.TestKit, c *C) {
 	}
 
 	// get c4 column id
-	testddlutil.SessionExecInGoroutine(s.store, "alter table t2 drop column c4", done)
+	testddlutil.SessionExecInGoroutine(s.store, "test_db", "alter table t2 drop column c4", done)
 
 	ticker := time.NewTicker(s.lease / 2)
 	defer ticker.Stop()
@@ -3839,7 +3022,7 @@ func (s *testDBSuite2) TestAddNotNullColumn(c *C) {
 	tk.MustExec("create table tnn (c1 int primary key auto_increment, c2 int)")
 	tk.MustExec("insert tnn (c2) values (0)" + strings.Repeat(",(0)", 99))
 	done := make(chan error, 1)
-	testddlutil.SessionExecInGoroutine(s.store, "alter table tnn add column c3 int not null default 3", done)
+	testddlutil.SessionExecInGoroutine(s.store, "test_db", "alter table tnn add column c3 int not null default 3", done)
 	updateCnt := 0
 out:
 	for {
@@ -4578,52 +3761,6 @@ func (s *testDBSuite4) TestIfExists(c *C) {
 	s.mustExec(tk, c, "alter table t2 drop partition if exists p1")
 	c.Assert(tk.Se.GetSessionVars().StmtCtx.WarningCount(), Equals, uint16(1))
 	tk.MustQuery("show warnings").Check(testutil.RowsWithSep("|", "Note|1507|Error in list of partitions to DROP"))
-}
-
-func testAddIndexForGeneratedColumn(tk *testkit.TestKit, s *testDBSuite5, c *C) {
-	tk.MustExec("use test_db")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(y year NOT NULL DEFAULT '2155')")
-	defer s.mustExec(tk, c, "drop table t;")
-	for i := 0; i < 50; i++ {
-		s.mustExec(tk, c, "insert into t values (?)", i)
-	}
-	tk.MustExec("insert into t values()")
-	tk.MustExec("ALTER TABLE t ADD COLUMN y1 year as (y + 2)")
-	_, err := tk.Exec("ALTER TABLE t ADD INDEX idx_y(y1)")
-	c.Assert(err, IsNil)
-
-	t := s.testGetTable(c, "t")
-	for _, idx := range t.Indices() {
-		c.Assert(strings.EqualFold(idx.Meta().Name.L, "idx_c2"), IsFalse)
-	}
-	// NOTE: this test case contains a bug, it should be uncommented after the bug is fixed.
-	// TODO: Fix bug https://github.com/pingcap/tidb/issues/12181
-	// s.mustExec(c, "delete from t where y = 2155")
-	// s.mustExec(c, "alter table t add index idx_y(y1)")
-	// s.mustExec(c, "alter table t drop index idx_y")
-
-	// Fix issue 9311.
-	tk.MustExec("drop table if exists gcai_table")
-	tk.MustExec("create table gcai_table (id int primary key);")
-	tk.MustExec("insert into gcai_table values(1);")
-	tk.MustExec("ALTER TABLE gcai_table ADD COLUMN d date DEFAULT '9999-12-31';")
-	tk.MustExec("ALTER TABLE gcai_table ADD COLUMN d1 date as (DATE_SUB(d, INTERVAL 31 DAY));")
-	tk.MustExec("ALTER TABLE gcai_table ADD INDEX idx(d1);")
-	tk.MustQuery("select * from gcai_table").Check(testkit.Rows("1 9999-12-31 9999-11-30"))
-	tk.MustQuery("select d1 from gcai_table use index(idx)").Check(testkit.Rows("9999-11-30"))
-	tk.MustExec("admin check table gcai_table")
-	// The column is PKIsHandle in generated column expression.
-	tk.MustExec("ALTER TABLE gcai_table ADD COLUMN id1 int as (id+5);")
-	tk.MustExec("ALTER TABLE gcai_table ADD INDEX idx1(id1);")
-	tk.MustQuery("select * from gcai_table").Check(testkit.Rows("1 9999-12-31 9999-11-30 6"))
-	tk.MustQuery("select id1 from gcai_table use index(idx1)").Check(testkit.Rows("6"))
-	tk.MustExec("admin check table gcai_table")
-}
-
-func (s *testDBSuite5) TestAddIndexForGeneratedColumn(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	testAddIndexForGeneratedColumn(tk, s, c)
 }
 
 func (s *testDBSuite5) TestModifyGeneratedColumn(c *C) {
@@ -6537,7 +5674,7 @@ func testDropIndexes(c *C, store kv.Storage, lease time.Duration, createSQL, dro
 	}
 	jobIDExt, resetHook := setupJobIDExtCallback(ctx)
 	defer resetHook()
-	testddlutil.SessionExecInGoroutine(store, dropIdxSQL, done)
+	testddlutil.SessionExecInGoroutine(store, "test_db", dropIdxSQL, done)
 
 	ticker := time.NewTicker(lease / 2)
 	defer ticker.Stop()
